pipeline/requirements.txt
anthropic
click
feedparser
firecrawl-py
pyyaml
requests
